+++ src/Runner/src/Agent/Interfaces/RunnerHandlerInterface.php
@@ -6,11 +6,7 @@ namespace PCIT\Runner\Agent\Interfaces;
 
 interface RunnerHandlerInterface
 {
-    public function handle(int $job_id): void
-    {
-    }
+    public function handle(int $job_id);
 
-    public function runStep(int $job_id, string $container_config, string $step = null): void
-    {
-    }
+    public function runStep(int $job_id, string $container_config, string $step = null);
 }